v0.1.0 EfaMonitor.DmCore.TransportService

the process responsible for calling the rest endpoint of the transport service. the process is especially calls the departure monitor endpoint and is not expected to call anyother api. the process works on 'async' calls, i.e., accepts only casts and will ignore any other messages

the type specs is as follows: request: {pid, dm_request_type,DepartureMonitorHttpRequest.t }

response: {dm_request_type, service_name, dm_response}

Types

dm_request_type()

dm_request_type() :: {:raw, :lines}

dm_response()

dm_response() ::
  {:ok, map()}
  | {:error, :retry}
  | {:error, {:station_not_found, list_of_suggestions :: [String.t()]}}
  | {:error, reason :: binary()}

service_name()

service_name() :: atom()

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link(list, arg)